Add authentication to project POST
